INTERLOK-3296 Restore MarshallingBaseCase
- This is used by xinclude
